[update] フロントエンドの実装 - jsのソースを機能ごとに分割した。
