+++ app/frontend/js/constants.js
@@ -1,0 +1,8 @@
+// バックエンドAPIのURL
+export const API_URL = "http://localhost:8080";
+
+// 役職のリスト
+export const positions = ["一般社員", "主任", "課長", "部長"];
+
+// 部署のリスト
+export const departments = ["技術部", "経理部", "人事部", "営業部"];
