+++ app/frontend/register.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>社員登録</title>
   <link rel="stylesheet" href="style.css">
-  <script defer src="app.js"></script>
+  <script type="module" src="js/main.js"></script>
 </head>
 <body id="register-page">
   <nav>
