+++ app/frontend/update.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>社員情報更新</title>
   <link rel="stylesheet" href="style.css">
-  <script defer src="app.js"></script>
+  <script type="module" src="js/main.js"></script>
 </head>
 <body id="update-page">
   <nav>
